backend/requirements.txt
# --- Core Web Server ---
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# --- AI & Forensics ---
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0

# --- Middleware & Utils ---
fastapi-cors==0.0.1
python-multipart==0.0.6
python-owasp-zap-v2.4==0.1.0
boto3==1.42.68

# --- Optional/Dev ---
pytest==7.4.3
httpx==0.25.1
